almost final
Всё работает. Проверил на ворнинги. осталось название файликов поменять.

+++ 0001.cpp
@@ -13,7 +13,7 @@ int main() {
 
         file >> n >> m;
 
-        for (int i = 0; i < m; i++) {
+        for (uint_fast64_t i = 0; i < m; i++) {
             file >> a >> b;
             graph.SetEdges(a, b, 0);
         }
